news_letter/main.py
from crewai import Crew, Process
from langchain_openai import ChatOpenAI
from agents import AINewsLetterAgents
from tasks import AINewsLetterTasks
from file_io import save_markdown
import os
from dotenv import load_dotenv

# Carregar variáveis de ambiente
load_dotenv()

# Inicializar o modelo GPT-4 do OpenAI
OpenAIGPT4 = ChatOpenAI(
        model="gpt-4"
    )

# Inicializar os agentes e tarefas
agents = AINewsLetterAgents()
tasks = AINewsLetterTasks()

editor = agents.editor_agent()
news_fetcher = agents.news_fetcher_agent()
news_analyzer = agents.news_analyzer_agent()
newsletter_compiler = agents.newsletter_compiler_agent()

fetch_news_task = tasks.fetch_news_task(news_fetcher)
analyze_news_task = tasks.analyze_news_task(news_analyzer, [fetch_news_task])
compile_newsletter_task = tasks.compile_newsletter_task(
    newsletter_compiler, [analyze_news_task], save_markdown)

# Formar a crew
crew = Crew(
    agents=[editor, news_fetcher, news_analyzer, newsletter_compiler],
    tasks=[fetch_news_task, analyze_news_task, compile_newsletter_task],
    process=Process.hierarchical,
    manager_llm=OpenAIGPT4,
    verbose=True
)

# Executar a crew
try:
    results = crew.kickoff()
    print("Crew Work Results:")
    print(results)
except Exception as e:
    print(f"Erro ao executar a crew: {e}")
