news_letter/requirements.txt
crewai
langchain_openai
python-dotenv
requests
